various adjustments and midi buffer
sampling frequency changed to 44100 Hz (tim1 period is now 1450).
this brings back the maximum 8 voices, which has now been defined to be
8.
note_off and note_on functions now reset data[0] and data[1].
midi data now goes into a ring buffer of size 128. previously, midi data
would get read incorrectly or missed as the midi process routine took too long, causing notes to get stuck. now the uart interrupt just puts data into the buffer, and the midi process routine is done in the main loop. this also means that the max voices needs to be valid because if max voices is too high, the synth routine won't exit fast enough to go back to the main loop and process midi messages. lowering sampling rate or max voices fixes this.
add_voice function now reactivates notes that are already in the array,
if the note was pressed twice. if more notes than max voices are
pressed, then the latest note who is in release state will be
overwritten with the new note.